Made the user profile page and sidebar component

+++ pet-adoption-api/src/main/java/petadoption/api/pet/Pet.java
@@ -19,7 +19,7 @@ public class Pet {
     private String petBreed;
     private String petGender;
     private int petAge;
-    private int petWeight;
+    private double petWeight;
     private String petSpecies;
     private String color;
 }
